requirements.txt
fastapi==0.116.0
uvicorn==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
boto3==1.39.4
langchain==0.3.26
langchain-aws==0.2.27
langgraph==0.5.3
langsmith==0.4.4
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
mangum==0.19.0
pytest==8.4.1
pytest-asyncio==1.0.0
langchain_community==0.3.27